+++ packages/core/rollup.config.js
@@ -1,22 +1,22 @@
-import typescript from 'rollup-plugin-typescript2'
-import { terser } from 'rollup-plugin-terser'
-import commonjs from 'rollup-plugin-commonjs'
-import resolve from 'rollup-plugin-node-resolve'
-import json from 'rollup-plugin-json'
-import pkg from './package.json'
+import typescript from 'rollup-plugin-typescript2';
+import { terser } from 'rollup-plugin-terser';
+import commonjs from 'rollup-plugin-commonjs';
+import resolve from 'rollup-plugin-node-resolve';
+import json from 'rollup-plugin-json';
+import pkg from './package.json';
 
 export default {
     input: 'src/index.ts',
     output: [
         {
             file: pkg.main,
-            name: "core",
+            name: 'core',
             format: 'umd',
             sourcemap: true
         },
         {
-            file: "./dist/core.umd.min.js",
-            name: "core.min",
+            file: './dist/core.umd.min.js',
+            name: 'core.min',
             format: 'umd',
             sourcemap: true
         },
@@ -50,10 +50,10 @@ export default {
         // // which external modules to include in the bundle
         // // https://github.com/rollup/rollup-plugin-node-resolve#usage
         resolve({
-            mainFields: ['module', 'main', "browser"]
+            mainFields: ['module', 'main', 'browser']
         }),
 
         // // Resolve source maps to the original source
         // sourceMaps(),
     ],
-}
+};