+++ packages/react-hooks/src/Glue.tsx
@@ -1,11 +1,10 @@
 import React, { createContext, memo } from "react";
 import { node, object, func } from "prop-types";
-import GlueWeb from "@glue42/web";
-import { Glue42Core } from "@glue42/core";
+import GlueWeb, { Glue42Web } from "@glue42/web";
 import { useGlueInit } from "./useGlueInit";
 import { GlueProviderProps } from "../react-hooks";
 
-export const GlueContext = createContext<Glue42Core.GlueCore>(null);
+export const GlueContext = createContext<Glue42Web.API>(null);
 
 export const GlueProvider: React.FC<GlueProviderProps> = memo(
     ({ children, fallback = null, config = {}, glueFactory = GlueWeb }) => {