+++ packages/react-hooks/src/useGlue.ts
@@ -1,9 +1,9 @@
 import { useContext, useEffect, useState } from "react";
 import { GlueContext } from "./Glue";
-import { Glue42Core } from "@glue42/core";
+import { Glue42Web } from "@glue42/web";
 
 export const useGlue = <T = undefined>(
-    cb: (glue: Glue42Core.GlueCore, ...dependencies: any[]) => void | Promise<T> | T,
+    cb: (glue: Glue42Web.API, ...dependencies: any[]) => void | Promise<T> | T,
     dependencies: any[] = []
 ): T => {
     const [result, setResult] = useState<T>();