+++ packages/core/changelog.md
@@ -1,2 +1,6 @@
+5.0.2
+feat: added in lerna release pipeline
+5.0.1
+fix: breaks if server returns null object ("server returns null result" test)
 5.0.0
 BREAKING CHANGE: removed support for Glue42 v2